Extract TestCaseEditor from the modal, and let suites live under an agent
Groundwork for rendering suites as a tree in KnowledgeExplorer.

All of the authoring UI — prompt box, expectations builder, judge rules, suite
and build pickers — was inside AddTestCaseModal. Showing a case in a right-hand
pane instead of a dialog would have meant a second copy of the expectations
builder, which is the bulk of those 1100 lines. It moves to TestCaseEditor,
which the modal now hosts as dialog chrome. The editor exposes save/runNow and
their pending state so a host renders its own footer rather than the editor
guessing at one, and emits `close` alongside `update:modelValue` so a pane can
listen without pretending to be a v-model.

Suites also become addressable per agent: TestSuiteCreate/Schema carry
data_source_id, and GET /tests/suites takes data_source_id (one agent's shelf)
or scope=global (the org-wide ones). Creating a suite homed on an agent takes
manage_evals on that agent; an org-wide one stays org-level.
